refactoring statement generation
