+++ generative_social_choice/queries/query_chatbot_personalization.py
@@ -10,6 +10,40 @@ from generative_social_choice.utils.gpt_wrapper import get_probabilities_from_co
 
 
 ##################################################
+# Classes for statement generation
+##################################################
+
+class SimplePersonalizationAgent(Agent):
+    """Simple agent representation which doesn't require connecting to any LLM
+    but can't be used to get approvals.
+    
+    We use this class in statement generation as computing new approvals
+    is unnecessary."""
+
+    def __init__(
+        self,
+        *,
+        id: str,
+        survey_responses: pd.DataFrame,
+        summary: str,
+    ):
+        self.id = id
+        self.survey_responses = survey_responses
+        self.summary = summary
+
+    def get_id(self):
+        return self.id
+
+    def get_description(self):
+        return self.summary
+
+    def get_approval(
+        self, statement: str, use_logprobs: bool = True
+    ) -> tuple[float, list[LLMLog]]:
+        raise NotImplementedError()
+
+
+##################################################
 # Discriminative queries
 ##################################################
 
@@ -176,7 +210,7 @@ class ChatbotPersonalizationGenerator(Generator):
             "model": model,
             "temperature": gpt_temperature,
             "max_tokens": 500,
-            "top_p": 1,
+            "top_p": 1,  # Does this make sense for temperature>0?
             "frequency_penalty": 0,
             "presence_penalty": 0,
             "stop": None,
